+++ nixos-modules/store.nix
@@ -61,7 +61,7 @@ let
     else { };
   activationHelperSourcePackage = pkgs.rustPlatform.buildRustPackage {
     pname = "d2b-activation-helper";
-    version = "0.0.0-bootstrap";
+    version = "2.0.0";
     src = packagesSrc;
     inherit cargoLock;
     cargoBuildFlags = [ "--package" "d2b-host" "--bin" "d2b-activation-helper" ];